+++ Airship/AirshipCore/Tests/AirshipContactTest.swift
@@ -25,6 +25,7 @@ class AirshipContactTest: XCTestCase {
         self.privacyManager = AirshipPrivacyManager(
             dataStore: self.dataStore,
             defaultEnabledFeatures: .all,
+            resetEnabledFeatures: false,
             notificationCenter: self.notificationCenter
         )