+++ Airship/AirshipCore/Tests/RemoteConfigManagerTest.swift
@@ -19,10 +19,10 @@ class RemoteConfigManagerTest: XCTestCase {
             dataStore: dataStore
         )
 
-
         self.privacyManager = AirshipPrivacyManager(
             dataStore: self.dataStore,
             defaultEnabledFeatures: .all,
+            resetEnabledFeatures: false,
             notificationCenter: self.notificationCenter
         )